% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Evaluation.R
\name{align_CellType}
\alias{align_CellType}
\title{Align the cell types from the prediction with the truth.}
\usage{
align_CellType(tt0)
}
\arguments{
\item{tt0}{a N*N table.}
}
\value{
the matched (re-ordered) table
}
\description{
Align the cell types from the prediction with the truth.
}
\examples{
vec1 = rep(1:4, each=100)
vec2 = sample(vec1)
tb = table(vec1, vec2)
#tb_arg = align_CellType(tb)
}
