% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_F.R
\name{cal_F2}
\alias{cal_F2}
\title{Calculate the gene-level F score and corresponding significance level.}
\usage{
cal_F2(Y, classes)
}
\arguments{
\item{Y}{A gene expression matrix}

\item{classes}{The initial cluster labels NA values are allowed. This can directly from the \code{Consensus} function.}
}
\value{
The score vector
}
\description{
Calculate the gene-level F score and corresponding significance level.
}
\examples{
data(Yan)
cal_F2(Y, classes = trueclass)
}
