% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_F.R
\name{cal_Fisher2}
\alias{cal_Fisher2}
\title{Calculate the gene-level fisher score.}
\usage{
cal_Fisher2(Y, classes)
}
\arguments{
\item{Y}{A gene expression matrix}

\item{classes}{The initial cluster labels NA values are allowed. This can directly from the \code{Consensus} function.}
}
\value{
The score vector
This is from the paper https://arxiv.org/pdf/1202.3725.pdf
Vector based calculation
}
\description{
Calculate the gene-level fisher score.
}
