% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Evaluation.R
\name{cal_metrics}
\alias{cal_metrics}
\title{Calculate 3 metrics and these methods are exported in C codes.
flag = 1 --- Rand index, flag = 2 --- Fowlkes and Mallows's index, flag = 3 --- Jaccard index}
\usage{
cal_metrics(cl1, cl2, randMethod = c("Rand", "FM", "Jaccard"))
}
\arguments{
\item{cl1}{a vector}

\item{cl2}{a vector}

\item{randMethod}{a string chosen from "Rand", "FM", or "Jaccard"}
}
\value{
a numeric vector including three values
}
\description{
Calculate 3 metrics and these methods are exported in C codes.
flag = 1 --- Rand index, flag = 2 --- Fowlkes and Mallows's index, flag = 3 --- Jaccard index
}
