% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{cal_MSE}
\alias{cal_MSE}
\title{Standard way to preprocess the count matrix. It is the QC step for the genes.}
\usage{
cal_MSE(Ynorm, cluster, return_mses = FALSE)
}
\arguments{
\item{Ynorm}{A normalized gene expression matrix. If not, we will normalize it for you.}

\item{cluster}{The clustering outcomes. Specifically, they are cluster labels.}

\item{return_mses}{True or False indicating whether returning the MSE.}
}
\value{
The MSE of the clustering centers with the predicted Y.
}
\description{
Standard way to preprocess the count matrix. It is the QC step for the genes.
}
\examples{
data(Yan)
Ynorm = Norm_Y(Y)
cluster = trueclass
MSE_res = cal_MSE(Ynorm, cluster)
}
