% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Evaluation.R
\name{eval_Cluster}
\alias{eval_Cluster}
\title{Calculate the a series of the evaluation statistics.}
\usage{
eval_Cluster(vec1, vec2)
}
\arguments{
\item{vec1}{a vector.}

\item{vec2}{a vector. x and y are with the same length.}
}
\value{
a vector of evaluation metrics
}
\description{
Calculate the a series of the evaluation statistics.
}
\examples{
vec2 = vec1 = rep(1:4, each = 100)
vec2[1:10] = 4
acc = eval_Cluster(vec1, vec2)
}
