% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEAST.R
\name{FEAST}
\alias{FEAST}
\title{FEAST main function}
\usage{
FEAST(
  Y,
  k = 2,
  num_pcs = 10,
  dim_reduce = c("pca", "svd"),
  split = FALSE,
  batch_size = 1000
)
}
\arguments{
\item{Y}{A expression matrix. Raw count matrix or normalized matrix.}

\item{k}{The number of input clusters (best guess).}

\item{num_pcs}{The number of top pcs that will be investigated through the consensus clustering.}

\item{dim_reduce}{dimension reduction methods chosen from pca or svd.}

\item{split}{boolean. If T, using subsampling to calculate the gene-level significance.}

\item{batch_size}{when split is true, need to claim the batch size for spliting the cells.}
}
\value{
the rankings of the gene-significance.
}
\description{
FEAST main function
}
\examples{
data(Yan)
k = length(unique(trueclass))
ixs = FEAST(Y, k=k)
}
