% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Load_CountMatrix}
\alias{Load_CountMatrix}
\title{Load count matrix}
\usage{
Load_CountMatrix(CountMatrix_path)
}
\arguments{
\item{CountMatrix_path}{A path to a tab-separated ASCII text counts file}
}
\description{
Load counts file - A tab-separated ASCII text file.
A matrix \eqn{C} of samples by taxa with the sources and sinks.
The row names are the unique sample ids. The column names are the unique taxa ids.
\eqn{Cij} is the read counts of taxa j in sample i each sample. Note that this order must be respected
}
