% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Load_metadata}
\alias{Load_metadata}
\title{Load_metadata}
\usage{
Load_metadata(metadata_path)
}
\arguments{
\item{metadata_path}{A path to a tab-separated ASCII text metadata file.}
}
\description{
Load metadata file - A tab-separated ASCII text file.
The row names of  metadata table are the unique sample ids.
The metadata file has three colunms (i.e., 'Env', 'SourceSink', 'id').
The first column is a description of the sampled environment (e.g., human gut), the second column indicates
if this sample is a source or a sink (can take the value 'Source' or 'Sink'). The fourth column is the Sink-Source id.
When using multiple sinks, each tested with the same group of sources, only the rows with 'SourceSink' = Sink will get an id
(between 1 - number of sinks in the data). In this scenatio, the sources ids are blank. When using multiple sinks,
each tested with a distinct group of sources, each combination of sink and its corresponding sources should get the
same id (between 1 - number of sinks in the data). Note that these names must be respected.
}
