% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{log_TPM_SC3}
\alias{log_TPM_SC3}
\title{Standard way to preprocess the gene expression data in SC3 method}
\usage{
log_TPM_SC3(Y)
}
\arguments{
\item{Y}{A gene expression data (Raw count matrix)}
}
\value{
A processed gene expression matrix. It is already log transformed and normalized by CPM
Followed by the step from SC3 to do the gene filtering.
> 2 counts and < 6%; expressed in more than 94% of the cells
}
\description{
Standard way to preprocess the gene expression data in SC3 method
}
