% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Norm_Y}
\alias{Norm_Y}
\title{Normalize the count expression matrix by the size factor and take the log transformation.}
\usage{
Norm_Y(Y)
}
\arguments{
\item{Y}{a count expression matrix}
}
\value{
a normalized matrix
}
\description{
Normalize the count expression matrix by the size factor and take the log transformation.
}
\examples{
data(Yan)
Ynorm = Norm_Y(Y)
}
