% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{PlotMultipleSinks}
\alias{PlotMultipleSinks}
\title{PlotMultipleSinks with different sources - use this plotting option to plot multiple sinks
each with a different set of sources.}
\usage{
PlotMultipleSinks(Filename, figures, dir_path)
}
\description{
PlotMultipleSinks with different sources - use this plotting option to plot multiple sinks
each with a different set of sources.
}
