% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{PlotPerSinkSourceContribution}
\alias{PlotPerSinkSourceContribution}
\title{Plot source contribution per sink}
\usage{
PlotPerSinkSourceContribution(SourceNames, source_contribution,
  sink_name = "Example", dir_path, save_plot)
}
\description{
Plot source contribution per sink
}
