% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{PlotSourceContribution}
\alias{PlotSourceContribution}
\title{PlotSourceContribution}
\usage{
PlotSourceContribution(SinkNames, SourceNames, Same_sources_flag,
  mixing_proportions, dir_path, Plot_title = "Example", N)
}
\arguments{
\item{SinkNames}{A vector with the sink names to plot.}

\item{SourceNames}{A vector with all the sources' names.}

\item{Same_sources_flag}{A boolian value indicating the source-sink plotting assignment.
Same_sources_flag = 1 if the same sources are assigned to the pre-defined sink samples , otherwise = 0.}

\item{mixing_proportions}{A list of vectors, where entry i corresponds to the vector of source
contributions (summing to 1) to sink i.}

\item{dir_path}{A path to an output .png file.}

\item{Plot_title}{Plot's title and output .png file's name.}

\item{N}{Number of barplot in each output .png file.}
}
\description{
Save a stacked barplot of source contributions to a group of pre-defined sink samples,
sharing the same sources.
}
\examples{
\donttest{

#Load metadata table

#Load count matrix

#Calculate the sources' contributions to
#each sink sample in the data
FEAST_output <- FEAST(count_matrix = otus, metadata = metadata,
                     different_sources_flag = 1)

#Plot and save a stacked barplot of source contributions
#to the first four sink samples
PlotSourceContribution(SinkNames, SourceNames, Same_sources_flag,
                      mixing_proportions, dir_path, Plot_title = "Example",N)

}

}
