% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{process_Y}
\alias{process_Y}
\title{Standard way to preprocess the count matrix. It is the QC step for the genes.}
\usage{
process_Y(Y, thre = 2)
}
\arguments{
\item{Y}{A gene expression data (Raw count matrix)}

\item{thre}{The threshold of minimum number of cells expressing a certain gene (default =2)}
}
\value{
A processed gene expression matrix. It is \emph{not log transformed}
}
\description{
Standard way to preprocess the count matrix. It is the QC step for the genes.
}
\examples{
data(Yan)
YY = process_Y(Y, thre=2)
}
