% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_Evaluation.R
\name{Purity}
\alias{Purity}
\title{Calculate the purity between two vectors.}
\usage{
Purity(x, y)
}
\arguments{
\item{x}{a vector.}

\item{y}{a vector. x and y are with the same length.}
}
\value{
the purity score
}
\description{
Calculate the purity between two vectors.
}
