% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Popular_Clust_Methods.R
\name{SC3_Clust}
\alias{SC3_Clust}
\title{SC3 Clustering}
\usage{
SC3_Clust(Y, k = NULL, input_markers = NULL)
}
\arguments{
\item{Y}{A expression matrix. It is recommended to use the raw count matrix.}

\item{k}{The number of clusters. If it is not provided, k is estimated by the default method in SC3.}

\item{input_markers}{A character vector including the featured genes. If they are not presented, SC3 will take care of this.}
}
\value{
the clustering labels and the featured genes.
}
\description{
SC3 Clustering
}
