% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{setUp_BPPARAM}
\alias{setUp_BPPARAM}
\title{set up for the parallel computing for biocParallel.}
\usage{
setUp_BPPARAM(nProc = 0, BPPARAM = NULL)
}
\arguments{
\item{nProc}{number of processors}

\item{BPPARAM}{bpparameter from bpparam}
}
\value{
BAPPARAM settings
}
\description{
This function sets up the environment for parallel computing.
}
\examples{
setUp_BPPARAM(nProc=1)
}
\keyword{internal}
