% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Popular_Clust_Methods.R
\name{Seurat_Clust}
\alias{Seurat_Clust}
\title{Seurat Clustering}
\usage{
Seurat_Clust(Y, knn = 10, resolution = 0.5, input_markers = NULL)
}
\arguments{
\item{Y}{A expression matrix. It is recommended to use the raw count matrix.}

\item{knn}{of nearest neighbors (knn).}

\item{resolution}{clustering resolution (0, 1) (default = 0.5).}

\item{input_markers}{A character vector including the featured genes. If they are not presented, SC3 will take care of this.}
}
\value{
the clustering labels and the featured genes.
input_markers is a character vector
}
\description{
Seurat Clustering
}
