\name{trueclass}
\alias{trueclass}
\docType{data}
\title{An example single cell dataset for the cell label information (Yan)}
\description{
  The true cell type labels for Yan dataset. It includes 8 different cell types.
}
\usage{data("Yan")}
\examples{
data("Yan")
table(trueclass)
}
\keyword{datasets}
