% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consensus_Clustering_Mclust.R
\name{vector2matrix}
\alias{vector2matrix}
\title{function for convert a vector to a binary matrix}
\usage{
vector2matrix(vec)
}
\arguments{
\item{vec}{a vector.}
}
\value{
a n by n binary matrix indicating the adjacency.
}
\description{
function for convert a vector to a binary matrix
}
