% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEAST.R
\name{FEAST}
\alias{FEAST}
\title{A scalable algorithm for quantifying the origins of complex microbial communities}
\usage{
FEAST(C, metadata, EM_iterations = 1000, COVERAGE = NULL,
  different_sources_flag = 1)
}
\arguments{
\item{C}{An \eqn{m} by \eqn{n} count matrix, where \eqn{m} is the number samples and \eqn{n} is the number of taxa.}

\item{metadata}{An \eqn{m} by \eqn{3} table, where \eqn{m} is the number samples.
The metadata table has three colunms (i.e., 'Env', 'SourceSink', 'id').
The first column is a description of the sampled environment (e.g., human gut), the second column indicates
if this sample is a source or a sink (can take the value 'Source' or 'Sink'). The fourth column is the Sink-Source id.
When using multiple sinks, each tested with the same group of sources, only the rows with 'SourceSink' = Sink will get an id
(between 1 - number of sinks in the data). In this scenatio, the sources ids are blank. When using multiple sinks,
each tested with a distinct group of sources, each combination of sink and its corresponding sources should get the
same id (between 1 - number of sinks in the data). Note that these names must be respected.}

\item{EM_iterations}{A numeric value indicating the number of EM iterations (default 1000).}

\item{COVERAGE}{A numeric value indicating the rarefaction depth (default = minimal sequencing depth within each group of sink
and its corresponding sources).}

\item{different_sources_flag}{A boolian value indicating the source-sink assignment.
different_sources_flag = 1 if different sources are assigned to each sink , otherwise = 0.}
}
\value{
P - an \eqn{S1} by \eqn{S2} matrix, where \eqn{S1} is the number sinks and \eqn{S2}
is the number of sources (including an unknown source). Each row in matrix \eqn{P} sums to 1.
\eqn{Pij} is the contribution of source j to sink i.
If \eqn{Pij} == NA it indicateds that source j was not used in the analysis of sink i.
}
\description{
FEAST performs microbial source tracking. As input, it takes
a count matrix \eqn{C}  of size \eqn{m} by \eqn{n} and a metadata table, of size \eqn{m} by \eqn{3},
where \eqn{m} is the number samples and \eqn{n} is the number of taxa.
The rownames of both \eqn{C} and the metadata table are the unique sample ids.
}
\examples{
\donttest{

#Load metadata table

#Load count matrix

#Calculate the sources' contributions to
#each sink sample in the data
FEAST_output <- FEAST(count_matrix = otus, metadata = metadata,
                     different_sources_flag = 1)
}

}
