% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual_result.R
\name{Visual_Rslt}
\alias{Visual_Rslt}
\title{Using clustering results based on feature selection to perform model selection.}
\usage{
Visual_Rslt(model_cv_res, trueclass)
}
\arguments{
\item{model_cv_res}{model selection result from \code{Select_Model_short_SC3}.}

\item{trueclass}{The real class labels}
}
\value{
a list of mse dataframe, clustering accuracy dataframe, and ggplot object.
}
\description{
Using clustering results based on feature selection to perform model selection.
}
\examples{
data(Yan)
k = length(unique(trueclass))
Y = process_Y(Y, thre = 2) # preprocess the data
set.seed(123)
rixs = sample(nrow(Y), 500)
cixs = sample(ncol(Y), 40)
Y = Y[rixs, ]
con_res = Consensus(Y, k=k)
# Not run
# mod_res = Select_Model_short_SC3(Y, cluster = con_res$cluster, top = c(100, 200))
library(ggpubr)
# Visual_Rslt(model_cv_res = mod_res, trueclass = trueclass)
}
