\name{Y}
\alias{Y}
\docType{data}
\title{An example single cell count expression matrix (Yan)}
\description{
  The data includes 124 cells about human preimplantation embryos and embryonic stem cells. It contains 19304 genes after removing genes with extreme high dropout rate.
}
\usage{data("Yan")}
\examples{
data("Yan")
Y[1:10, 1:4]
}
\keyword{datasets}
